Fixed bug due to divide by zero in case of scenes without participating media resulting in scaling by  inf
Also removed warnings

+++ Integrators/Progressive Photon Mapping/PPPMTracer.h
@@ -98,8 +98,8 @@ public:
 	}
 	void getPhotonsEmittedLastPass(float& nSurface, float& nVolume)
 	{
-		nSurface = m_uPhotonEmittedPassSurface;
-		nVolume = m_uPhotonEmittedPassVolume;
+		nSurface = (float)m_uPhotonEmittedPassSurface;
+		nVolume = (float)m_uPhotonEmittedPassVolume;
 	}
 
 	std::map<std::string, boost::variant<int, float>> getPixelInfo(int x, int y) const;
